requirements.txt
playwright
beautifulsoup4
patchright
streamlit
langchain
langchain-community
langchain-text-splitters
python-dotenv
html2text
streamlit-tags
langchain-groq
pandas
